Add sudo requirement

+++ run.sh
@@ -1,3 +1,8 @@
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root or sudo" 1>&2
+   exit 1
+fi
+
 echo "Stopping and removing existing proxy"
 docker stop socksvpn
 docker rm socksvpn
